simple_local_code_reviewer/requirements.txt
requests
huggingface-hub
llama-cpp-python
